refactor: rename Select -> MatchSelect, move equal type for MatchSelect to types.js
